refactor QA's page routing
